backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
sqlalchemy>=2.0,<3.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
